pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["evrik163 <evrik163@mail.ru>"]
readme = "README.md"
packages = [
  { include = "logic" },
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
asciinema = "^2.2.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
PyYAML = "^6.0"

[tool.poetry.scripts]
gendiff = "logic.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
